include only other user in ephemeral message payload
